Preview

+++ Fonts Downloader/CssFile.cs
@@ -17,23 +17,18 @@ namespace Fonts_Downloader
             { "900", "Black" },
         };
 
-        public void CreateCSS(IEnumerable<string> variants, string folderName, string fontName, bool minify=false, IEnumerable<string> subsets = null)
+        public void CreateCSS(IEnumerable<string> variants, string folderName, string fontName, bool woff, bool minify = false, IEnumerable<string> subsets = null)
         {
-            if (variants == null || !variants.Any() || string.IsNullOrEmpty(fontName) || string.IsNullOrEmpty(folderName))
-            {
-                string missingParameters = string.Join(", ",
-                    (variants == null || !variants.Any()) ? "Items" : "",
-                    string.IsNullOrEmpty(fontName) ? "SelectedFont" : "",
-                    string.IsNullOrEmpty(folderName) ? "FolderName" : "");
-                throw new ArgumentException($"One or more required parameters are missing or invalid: {missingParameters}");
-            }
 
-            var cssList = GenerateCssList(variants, fontName, subsets);
+            var cssList = GenerateCssList(variants, fontName, woff, subsets);
 
             if (cssList.Any())
             {
                 string fontFolder = Path.Combine(folderName, fontName.Replace(" ", ""));
-                Directory.CreateDirectory(fontFolder);
+                if (!Directory.Exists(fontFolder))
+                {
+                    Directory.CreateDirectory(fontFolder);
+                }
 
                 string cssFilePath = Path.Combine(fontFolder, $"{fontName.Replace(" ", "")}{(minify ? ".min" : "")}.css");
                 string cssContent = string.Join("\n", cssList);
@@ -43,12 +38,31 @@ namespace Fonts_Downloader
                     var minifier = new Minifier();
                     cssContent = minifier.MinifyStyleSheet(cssContent);
                 }
+                if (!File.Exists(cssFilePath) || (File.Exists(cssFilePath) && string.IsNullOrEmpty(File.ReadAllText(cssFilePath))))
+                    File.WriteAllText(cssFilePath, cssContent);
 
-                File.WriteAllText(cssFilePath, cssContent);
+                else
+                {
+                    string fileContents = File.ReadAllText(cssFilePath);  
+                    if (woff && fileContents.Contains(".ttf") || !woff && fileContents.Contains(".woff2"))
+                    {
+                        File.WriteAllText(cssFilePath, cssContent);
+                    }
+                    else
+                    {
+                        if (!fileContents.Contains(cssContent))
+                        {
+                            string newString = cssContent.Replace(fileContents, "");
+                            int index = fileContents.IndexOf(cssContent);
+                            if (index == -1)
+                                File.AppendAllText(cssFilePath, newString);
+                        }
+                    }
+                }
             }
         }
 
-        private List<string> GenerateCssList(IEnumerable<string> variants, string fontName, IEnumerable<string> subsets = null)
+        private List<string> GenerateCssList(IEnumerable<string> variants, string fontName, bool woff, IEnumerable<string> subsets = null)
         {
             var cssList = new List<string>();
 
@@ -57,17 +71,17 @@ namespace Fonts_Downloader
                 if (ParseCheckedItem(variant, out var fontStyle, out var fontWeight) &&
                     FontFileStyles.TryGetValue(fontWeight, out var fontFileStyle))
                 {
-                    if (subsets != null && (subsets.Any() || subsets != null))
+                    if (subsets != null && subsets.Any())
                     {
                         foreach (var subset in subsets)
                         {
-                            var css = GenerateFontFaceCss(fontName, fontStyle, fontWeight, fontFileStyle, subset.ToLower());
+                            var css = GenerateFontFaceCss(fontName, fontStyle, fontWeight, woff, fontFileStyle, subset.ToLower());
                             cssList.Add(css);
                         }
                     }
                     else
                     {
-                        var css = GenerateFontFaceCss(fontName, fontStyle, fontWeight, fontFileStyle);
+                        var css = GenerateFontFaceCss(fontName, fontStyle, fontWeight, woff, fontFileStyle);
                         cssList.Add(css);
                     }
 
@@ -84,7 +98,36 @@ namespace Fonts_Downloader
             return !string.IsNullOrWhiteSpace(fontWeight);
         }
 
-        private string GenerateFontFaceCss(string fontName, string fontStyle, string fontWeight, string fontFileStyle=null, string subset = null)
+        private string GenerateFontFaceCss(string fontName, string fontStyle, string fontWeight, bool woff, string fontFileStyle = null, string subset = null)
+        {
+            if (!ValidateFontParameters(fontName, fontStyle, fontWeight)) return string.Empty;
+
+            var subsetComment = !string.IsNullOrEmpty(subset) ? $"/*{subset}*/\n" : "";
+
+            var fontFileName = FontFileName(fontName, fontStyle, fontFileStyle);
+
+            var format = woff ? "woff2" : "ttf";
+            var formatAttribute = format == "ttf" ? "" : $" format('{format}')";
+
+            var css = $"{subsetComment}@font-face {{\n" +
+                       $"  font-family: '{fontName}';\n" +
+                       $"  font-style: {fontStyle};\n" +
+                       $"  font-weight: {fontWeight};\n" +
+                       "  font-display: swap;\n" +
+                       $"  font-stretch: 100%;\n" +
+                       $"  src: url('{fontFileName}.{format}'){formatAttribute};\n";
+
+            if (woff)
+            {
+                css += "  unicode-range: U+0460-052F, U+1C80-1C88, U+20B4, U+2DE0-2DFF, U+A640-A69F, U+FE2E-FE2F;\n";
+            }
+
+            css += "}";
+
+            return css;
+        }
+
+        private bool ValidateFontParameters(string fontName, string fontStyle, string fontWeight)
         {
             if (string.IsNullOrEmpty(fontName) || string.IsNullOrEmpty(fontStyle) || string.IsNullOrEmpty(fontWeight))
             {
@@ -94,23 +137,8 @@ namespace Fonts_Downloader
                     string.IsNullOrEmpty(fontWeight) ? "Font's weight" : "");
                 throw new ArgumentException($"One or more required parameters are missing or invalid: {missingParameters}");
             }
-            var subsetComment = !string.IsNullOrEmpty(subset) ? $"/*{subset}*/\n" : "";
-            return string.IsNullOrEmpty(subsetComment)
-       ? $"@font-face {{\n" +
-         $"font-family: '{fontName}';\n" +
-         $"font-style: {fontStyle};\n" +
-         $"font-weight: {fontWeight};\n" +
-         $"font-stretch: 100%;\n" +
-         $"src: url('{FontFileName(fontName, fontStyle, fontFileStyle)}.ttf')\n}}"
-       : $"{subsetComment}" +
-         $"@font-face {{\n" +
-         $"font-family: '{fontName}';\n" +
-         $"font-style: {fontStyle};\n" +
-         $"font-weight: {fontWeight};\n" +
-         $"font-stretch: 100%;\n" +
-         $"src: url('{FontFileName(fontName, fontStyle, fontFileStyle)}.ttf')\n}}";
-        
-    }
+            return true;
+        }
 
         private string FontFileName(string fontName, string fontStyle, string fontWeight)
         {
